uibuild

+++ controllers/journals.js
@@ -1,4 +1,5 @@
 const Journal = require("../models/journal");
+const User = require("../models/users");
 
 const journalControllers = {
   getAllJournals: async (req, res) => {
@@ -16,32 +17,47 @@ const journalControllers = {
   },
   addNewJournal: async (req, res) => {
     const { title, content, challenges } = req.body;
-    const existingTitle = await Journal.findOne({ title });
+    const user = await User.findById(req.user.id);
+    const existingTitle = await Journal.findOne({ title, user: user._id });
+
     if (existingTitle) {
       return res
         .status(400)
         .json({ error: `Journal with the title: ${title} already exists !` });
     }
-    const newJournal = new Journal({
+    const journal = new Journal({
       title,
       content,
       challenges,
+      user: user._id,
     });
-    await newJournal.save();
+    const newJournal = await journal.save();
+    user.journals = user.journals.concat(newJournal._id);
+    await user.save();
     res.status(201).send(newJournal);
   },
   deleteJournalById: async (req, res) => {
     const { id } = req.params;
-    const deletedJournal = await Journal.findByIdAndDelete(id);
-    if (!deletedJournal) {
-      return res.status(400).json({ error: "No item found" });
-    } else {
-      res.status(204).end();
+    const journal = await Journal.findById(id);
+    if (!journal) {
+      return res.status(404).json({ error: "No Journal found" });
     }
+    if (journal.user.toString() !== req.user.id) {
+      return res.status(400).json({ error: "Unauthorized Action" });
+    }
+    await Journal.findByIdAndDelete(id);
+    res.status(204).end();
   },
   updateJournalById: async (req, res) => {
     const { id } = req.params;
     const { title, content, challenges } = req.body;
+    const journal = await Journal.findById(id);
+    if (!journal) {
+      return res.status(404).json({ error: "No Journal Found" });
+    }
+    if (journal.user.toString() !== req.user.id) {
+      return res.status(401).json({ error: "Unauthorized Operation" });
+    }
     const updatedJournal = await Journal.findByIdAndUpdate(
       id,
       {
@@ -51,12 +67,17 @@ const journalControllers = {
       },
       { new: true, runValidators: true, context: "query" }
     );
+    res.status(200).json(updatedJournal);
+  },
 
-    if (!updatedJournal) {
-      return res.status(400).json({ error: "Invalid Id" });
-    } else {
-      res.status(200).json(updatedJournal);
+  getJournalsOfAUser: async (req, res) => {
+    const { userId } = req.params;
+    const user = await User.findById(userId);
+    if (!user) {
+      return res.status(404).json({ error: "No User found on this id" });
     }
+    const journals = await Journal.find({ user: userId });
+    res.status(200).json(journals);
   },
 };
